Make adding a function on circuit open easier
Instead of only being able to configure the callback upon service
creation, allow to configure it elsewhere. Every time #on_circuit_open
is called it will be done so with the circuit as an argument:

    circuit = Disyuntor.new(threshold: 3, timeout: 5)

    circuit.on_circuit_open do |c|
      "Circuit was open at #{c.last_open}"
    end

As before the value of the block will be returned as the value of
Disyuntor#try when it fails.