Circuit open callback fails by default
Previously it was possible to pass a block to #initialize to define a
custom circuit open callback upon initialization. This wasn't a
documented feature and it could be a source of unexpected bugs. This
change removes that possibility and defines the callback as a simple
`fail CircuitOpenError`. This will be the default action unless the user
defines a new callback by calling #on_circuit_open.