Ensure exactly one #try executes while timed out
When the circuit is open and the time out period has passed a call
to #try executes the block as a trial. This is intended to happen
only once. Currently this can happen more than once when (1) a
trial is "in progress" in a background thread.

The reason is that `#timed_out?` checks `@opened_at`, which isn't
updated until the trial finishes. `#timed_out?` will always be
`true` while a trial is in progress.

Calling `#open!` before performing the trial ensures a second call
to `#try` will not result in a trial since `#timed_out` is no longer
`true`.